requirements.txt
# external requirements
flask
flask-cors
cianparser
scikit-learn
pandas
boto3
python-dotenv
PyGithub
joblib
requests
numpy
catboost